Flush Errlmanager Queue Before Fetching Gard Records
When an error occurs or is injected into an EX target,
the gard record for the target does not make it into
the GUARD partition in time for the gard processing
in istep 6.9 because the associated error log is sent
asynchronously. This change flushes the errlmanager
message queue right before the processing of the
gard records to insure all gard records make it into
GUARD partition.

Change-Id: I06625672df7e926d6d080673241ad2ee4d87a7c9
CQ: SW461565
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75847
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75913