Workaround unused var in certain configurations
If we are not doing CHECKSTOP_ANALYSIS and we are in AXONE_BRING_UP
the we will not use the l_errl variable in host_start_stop_engine.
This causes the compile to fail because we elevate warnings to
errors. By casting the variable to void in all cases, this will
workaround the warning.

Change-Id: Ice3c477e2597ba29d5caf987e1d0d207617241c7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75671
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Luis P. Fernandez <luis.fernandez@ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>