Correct MSS accessor generation for Axone compiles
There is still some changes coming through for the p9a_mss_freq HWP.
When we tried to pull these changes in it exposed some problems we
had in how we were generating the accessor functions that the mss
hwps rely on. We were missing some XMLs and when we pulled them
in it caused collisions with a mirror of the accessor header we
pulled in by accident. This commit resolves the accessor issues.

Change-Id: Ia134ffba6e82eb93f9c0762dcd851866d4475de8
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75779
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>