Infer presense detection of Explorer chip from VPD EEPROM
During presense detection the explorer chip is being held in reset. In
this state it will not respond to any I2C operations and therefore we
cannot physically detect the chip during normal presense detection.
Instead we infer its presense based on the presense of the associated
VPD EEPROM.

Change-Id: Id07eb77fc9c3eab09e852b191ad4a32832880c38
RTC:203722
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76322
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M. Crowell <dcrowell@us.ibm.com>