Update OCMB 9-15 to have valid i2c and eeprom info
The simics model only has 9 valid ocmbs represented on the master
processor. ocmbs 0-7 are behind a 1-8 MUX and ocmb 8 is directly
connected. This leaves ocmbs 9-15 for us to fill out. The information
must be valid enough to allow the code to process the targets correctly,
but we must fake out some information for the the sake of the awkward
simics model. We have decided that for ocmbs 9-11 we will match
everything from ocmb8 except increment the devAddr of the I2C info
attributes A2,A4,A6, D2,D4,D6. For ocmbs 12-15 we have picked a new port
(2) and used the same dev addr increments. This slightly invalid data
allows the code to have the targets show up as PRESENT but not
FUNCTIONAL

Change-Id: I3aec520a04e89829554c277a4cf02e1981b7ed84
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/75999
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>