Gemini vs Explorer Presence Detection via SPD
Since the OCMB chip is held in reset until after presence detection the
IDEC register cannot be read to differentiate between Gemini and
Explorer chip types. To work around this issue, during the early part of
IPL when presence detection is occurring the OCMB IDEC function will instead
read the SPD and populate the necessary attributes with what is found
there. That will be used to determine the difference between Gemini and
Explorer until later when the OCMB IDEC register can be read from. At
that point the IDEC read will be executed again and the data read from
the OCMB IDEC register will be used to cross-check the data read from
the SPD. Any discrepancies will be handled with predictive error logs.

Change-Id: Ica664b06ff3488f48253d3ef02eff2d49c5d240d
RTC: 208696
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76108
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>