Compile and call exp_draminit_mc
Per the P9A IPL Flow documentation we must call exp_draminit_mc during
istep 13.12. On hardware this requires exp_draminit to have been called.
But in simulation it shouldnt matter. I make this note because at the
time of writing the the exp_draminit HWP is not being called.

RTC: 195556
Change-Id: I2db07c461cf9f10b85d51c091964b62574abb62e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76256
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Glenn Miles <milesg@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>