Separate communication data padding from endianess correcting code
In the OCBM inband communication code the func correctMMIOEndianForStruct
will flip around some of the struct's bytes to satisfy the endianess
changes needed between the OCMB and the P9A processor. This code was
also padding out the communication data so it was always an even multiple
of 8 bytes. This padding was getting skipped if the attribute was set
to skip the endianess correction. To fix this I decided to remove the
padding code from the endianess correction code and make a new function
that is being called before the correctMMIOEndianForStruct function.

Change-Id: I75007697e69d7923ea5f8378524e30f4dc589108
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76350
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: ANDRE A. MARIN <aamarin@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76362
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>