Only display ERRL user details callouts/strings when dumping error to console
This commit fixes a bug wherein, when Hostboot dumped an error log to the
console that contained user details with the same subsection ID as an ERRL
string but for a different component ID, the information, which was often in
binary format, would appear garbled, and would additionally terminate the BMC's
SOL session due to invalid characters.  The change narrows the scope of what
gets emitted to just ERRL component strings/callouts.

Change-Id: Ifbe83448a56955d44f0e775e123ffb95a6330b8a
CQ: SW463437
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76392
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>