Increase size allocated for HBI in axone pnor layout xml
Developers have been hitting size contraints when trying to build
a pnor image with test cases loaded. Recently we removed the CVPD
and DJVPD sections from the axone pnor layout so we have some
extra room to expand out the HBI image. This commit increases the size
of the HBI section from 16->20 MB. Some of that space is used
for ECC, so in total we get an extra 3.5 MB of codespace. My local
test shows that we are at 88.5% of the size utilization now for HBI, this
is down from the 115% which is was showing before. More work needs
to be done to cut down the size as more code is coming.

Change-Id: Ia012791220def9a324e29c7ae9dcb7ef85ede0c7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76501
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M. Crowell <dcrowell@us.ibm.com>