Remove workaround filtering invalid filenames
When Axone HWPs first started coming in from the memory team they
were naming some of the files the same name as what they did for
the nimbus version of these files. This caused issues when we
tried to compile because there were naming collisions. These
file names have changed to have the model prepended to the filename

Change-Id: I9868e39cb70c34f9eb3fbd7244b7c2065620f3eb
RTC: 207832
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76655
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M. Crowell <dcrowell@us.ibm.com>