Remove workaround setting EFF_DIMM_SIZE attribute
We were missing some functionality in the eff_config HWP that set
this attribute so we had to put a workaround in place to make sure
it was set correctly. This value should be getting read out of the
SPD during eff_config now so we can remove this workaround.

Change-Id: I373f92c0e2b2238dfac54d9ae50973ed3ff00a57
RTC: 207850
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76796
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>