Add call to setup derived attributes in p9a_mss_eff_config
Previously there was some work done to write the functions that
do the calculations for these derived attributes but the code
was not being called, and one of the functions had an incorrect
parameter. This commit addresses both of those issues and also
improves on a trace in eff_grouping that made debugging this issue
tricky.

Change-Id: I8394f1dcc29374afc765668fce79e55f9f823e7d
RTC: 207850
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76616
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: ANDRE A. MARIN <aamarin@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76661
Reviewed-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Christian R. Geddes <crgeddes@us.ibm.com>