Update simbuild for axone simics bringup
The XML for the RAM1 register was not being parsed correctly
resulting in too few registers being allocated in
uchip_regs.chip not defining all of the registers.  This
latest build adds those registers manually until the parser
can be fixed.  This build also sets the POR values for the
RAM1 registers.

Also changes OCMB I2C addresses to 0x40

Change-Id: Icd2df80874200741d82fc152cb4b8bdbc75c5bed
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76764
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>