P10: Support for Variable Cache Memory Size
This commit introduces support for variable initial size of cache
memory available to hostboot to run from. SBE will tell hostboot
how much cache memory is available for it to run from via the
SBE to bootloader communication area. This amount will be a multiple
of 4MB and will be at least 8MB. The cache size is propagated from
the bootloader to the HB kernel via BL to HB comm area. Note that
we will not perform half- or full-cache expansion anymore, since
the entire cache will be available right away.

Change-Id: I571fdf29300eed179846ab1e49248afb732ac604
RTC: 2089792
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78104
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>