Add logic to compute MSS_MRW_NVDIMM_SLOT_POSITION for ZZ
There are a variety of ways to number dimms.  For NVDIMMs there
is a system property that defines which positions we are allowed
to install the dimms in (ATTR_MSS_MRW_NVDIMM_PLUG_RULES).  Because
of the inconsistent numbering schemes we have explicitly defined
a new attribute on the dimm that corresponds to the system plug
rules.

Change-Id: If935e29eb108a6a922d8c1f7c69fdc12580c5958
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/77239
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>