Refactor keyword support for various DIMM types
The existing keyword logic didn't use bitmasking when searching for the
correct keyword entry. This commit refactors the code to allow for
bitmasking and changing the NA module specific keyword to mean that no
keyword was found for the given target rather than its previous
ambiguous meaning that could be confused with the ALL module specific
keyword.

Change-Id: I661b70c4eff2740911cd63f8c1042ee8a084d63a
RTC:203788
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/77357
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Glenn Miles <milesg@ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>