Clean up some comments and refactor how we determine new cache entry
While code was added to handle detecting new OCMB parts a new check
was added to make sure an entry was newly added to cache to make it so
we didnt do new part validation on entries that had not been cached yet.
This commit removes what was added and leverages existing information we
had earlier on in the caching algorith. Also in this commit a few minor
comments were addressed related to tracing and doxygen comments.

Change-Id: Ie3454181449320a6a570c15692cc1f71c819ba22
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/77814
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/80649
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>