Allow to set image id string via environment
Hostboot's Image Id is the first text string that a customer sees
when the host starts booting.
By default, image id is the version of hostboot itself. This
patch allows to change it into the more meaningful value, such as
a machine name, release version number of firmware and so on.

Custom image id string can be specified in the environment variable
HOSTBOOT_IMAGEID. If this variable is not defined, a build script
uses HOSTBOOT_P8_VERSION value. HOSTBOOT_VERSION removed because
it is not defined for P8 builds.

Resolves #177

Signed-off-by: Artem Senichev <a.senichev@yadro.com>
Change-Id: Iee8b46c08bd88d70bea510c60c8c9c4d17722986
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/77922
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>