Fixes MCBIST read bug in i2c mode
I2c reads are 32 bits. This means to read an IBM
logic register, the register must be read 2x, once for
the left half and once for the right half of the register.
This causes the read buffer to auto-increment 2x per read
instead of 1x per read.
As such, we have disabled the auto-increment feature in explorer
for all MCBISt tests.

JIRA EDBC-444

Change-Id: I21157376717177a02788ed7faeb76688df816727
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78347
Reviewed-by: Mark Pizzutillo <mark.pizzutillo@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: RYAN P. KING <rpking@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78352
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>