Update non-present dimm/ocmb i2c attributes
The dimm/ocmb i2c information should match up, especially for the
VPD_PRIMARY attribute. Also we had a problem where the settings we
had for the dimm target was getting targets that were not present
in the simics model showing up present to hostboot. By forcing targets
we no should not show up to have invalid devAddr fields set in the
attribute, we avoid having targets show up when they should not.

Change-Id: I3002e81d39ec78f3ba25c22ce9d29e718ee68df1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78415
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>