Correctly form LX record/keyword for OpenPOWER
Current OpenPOWER code doesn't properly form
the LX VPD record and instead tacks it into the
VINI record.  This causes PowerVM to fail to boot.
Correct this by fully forming a LXR0 record, populate
the LX keyword from the value in the system
MRW, and then append the fully formed LXR0 record to
the HDAT VPD.

Change-Id: I2128e4ddef8e3c5c1373fd4c6b5bf6de50901ef1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/80474
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>