Add interface to mark eeprom caches valid/invalid
There are certain scenarios where we might want to update the
table of contents at the start of the EECACHE section in pnor to
indicate we have decided that the cached copy of one of the eeproms
is no longer valid. This commit adds the ability to mark a cache
entry as either valid or invalid by calling setIsValidCacheEntry
in the EEPROM namespace and tests to verify this works. Future commits
will add this call to places it is needed.

Change-Id: I6b7dc807b7dbcaca3efa298790f862bc2c5f9ec6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78494
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>