Trace message if unable to understand its contents of eeprom cache
While checking if PN/SN has changed for a given part we will lookup
those PN/SN in the eeprom cache. If we are unable to understand the
contents of the eeprom cache then we want to make sure some sort of
message gets out to the user. This would likely happen in bringup or
lab scenario, and there are other logs generated that would lead you
to this code so this trace should suffice.

Change-Id: Id45c6de42ee6792e8d4bbfdc460da371a5bcb99e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78504
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/81699
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>