Add functions to back ATTR_FREQ_MCA_MHZ and ATTR_MC_PLL_BUCKET in P9A
The FREQ_MCA_MHZ is being used differently in Axone than previously
in P9 systems. Now this attribute is now tied in a relationship with
the MC_PLL_BUCKET and OMI_PLL_VCO attributes which is defined in the
OMI_PLL_FREQ_LIST in the EKB. Hostboot will use the FREQ_OMI_MHZ,
which was determined from the parsing DDIMM's VPD, to tell if we
have decided that a different PLL bucket than what we booted with
is more optimal. At least one scominit procedure is also looking up
these values. This commit provides that functions that will allow
the Hostboot platform to easily lookup these values.

Change-Id: I8e8d970c682d11a73f9fd1962782b613f55bee48
RTC: 207596
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78542
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83025
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>