Update bbuild and point axone simics at scandef.abr from bbuild
We had issues updating the backing build because the SBE code was
failing to boot without getting updated scandef information. Simics
was lagging behind updating their copy of the scandef so we will
just point directly to the one in the bbuild we want to use in the
cmd line call.

Change-Id: I258ccaecc3e17e82de31ff2d6d4359dcb92045b0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78701
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>