Add Enabling of Mambo Core to startup.simics
This commit adds a hack to enable mambo core to run instructions on
it in startup.simics script. This command was a part of entry.simics
script, but in order to allow both hostboot and SBE to have their own
set of hacks, the command was moved into the respective startup.simics
scripts.

Change-Id: Ieb5fa8686385fdb6df03e430c80fce23b8dbd2cf
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78972
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>