Clean up the attribute error user details
Moved the structural parts of the errludattribute code into source
files, instead of being generated.  Now we only pull in the
generated code where it is actually useful.  This makes updates
simpler and faster to compile.

Added a couple more attributes to the supported list of displayable
values.

Improved an error path to get more useful information from
unsupported attributes.

Change-Id: Ie339e4a1b0a6c8a1d33def92a76e36ba81bab011
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/79494
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>