Unit test for the firmware request sendAttributes
Added a unit test for the firmware request sendAttributes.  The unit
test exercises and verifies the correctness of the sendAttribute
firmware request method.

Did some massive cleanup of the testing framework/harness.  It was
becoming unwieldy and cumbersome to update.  I create methods for the
individual firmware request test instead of having the test reside in
over bloated method.

Added a section, in the README file, on how to test new interfaces
that are added for firmware request lineup.

Change-Id: I0ff7c0e48211b12d81be1f179ad3ee08e4d254af
RTC: 208343
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78934
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Glenn Miles <milesg@ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>