P10 Partial Good (commit 1/2)
This commit updates the partial-good rules for P10. Rules for
new targets are added, for old have been removed, and the test suite
was refactored and updated to reflect changes in the ruleset.

This commit does not include bug fixes facilitated by Simics testing;
those changes will be added in a later commit.

Change-Id: I071de6fc7c233ba7c8a37dcce771cbc7220f1466
RTC: 208782
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/79142
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Glenn Miles <milesg@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>