Start simics in vexec for Axone by default
Previously when you called 'hb startsimics' with the axone config
we were not running in a vexec shell. This was causing the pool
machines to crater if people were running 1 or 2 simics sessions.
This commit will check if the '-nre' flag has been passed in, or
if we are in a vexec shell already. If niether of those are true
then we will boot simics in a vexec shell, if they are -nre has
been passed or we are already inside a vexec then we will not boot
simics inside a vexec.

Change-Id: I17895bb109f18bbcc4b0daafe6c864531bf5994a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/79573
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>