Push ATTR_NVDIMM_ARMED down to FSP to handle MPIPL case
If the NVDIMM is not armed when the save/flush call is made during
a mpipl reboot, it will cause errors in the memory.  To avoid this
we need HBRT to push down the armed state to the FSP so that HWSV
can skip the save call in the case where the NVDIMM is not armed.

Change-Id: I16611e8fcb5274c3cb0aee6fec1300a95579b11b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89598
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>