Add optional params to build-script to skips some steps
build-script not only compiles Hostboot and primes the ODE sandbox
it also performs a copyright check and runs cxx tests. There is a
desire to run this script to generate a populated sandbox but we
dont need to run the copyright check or test cases so we need flags
to avoid that.

Change-Id: I93b698a0b8d634bedba43e25361771a0e08875fa
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/85173
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>