Temporary fix: skipping over printing doubles to avoid a crash
The printing of double, for the trace FAPI_INF, was never implemented
and thus when a double is encountered via open power, a crash would
happen.  This fix just simply skips over the double, printing nothing,
to avoid the crash.  An implementation of the double will follow
later.

This fix is done because Romulus is blocked on this issue.

Change-Id: I48804de3cfecc966ba9541e1981aca342d7cbb83
CQ: SW464805
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/79885
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>